requirements.txt
Flask
google-generativeai
google-cloud-aiplatform
Pillow
gunicorn